scrapers/requirements.txt
# Core Scrapy framework - Updated to patch all security vulnerabilities
Scrapy==2.13.4

# HTTP client with better SSL support
requests==2.31.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Azure Data Lake integration
azure-storage-blob==12.19.0
azure-identity==1.15.0

# FastAPI for scraper API - Updated to patch ReDoS vulnerability
fastapi==0.109.1
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Database connectivity
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Data validation and parsing
python-dateutil==2.8.2
validators==0.22.0

# Logging and monitoring
python-json-logger==2.0.7

# Anti-blocking tools
scrapy-user-agents==0.1.1
scrapy-rotating-proxies==0.6.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Environment management
python-dotenv==1.0.0
